refactor : ai model-domain 분리
